part update

+++ models/cal_evt.js
@@ -12,7 +12,7 @@ module.exports = sequelize.define('cal_evt', {
     evt_name: Sequelize.STRING,
     color: Sequelize.STRING,
     description: Sequelize.STRING,
-    score: Sequelize.INTEGER,
+    score: Sequelize.FLOAT,
     create_time: Sequelize.BIGINT,
     update_time: Sequelize.BIGINT,
     delete_time: Sequelize.BIGINT,
