+++ models/cal_evt_template.js
@@ -11,7 +11,7 @@ module.exports = sequelize.define('cal_evt_template', {
     evt_name: Sequelize.STRING,
     color: Sequelize.STRING,
     description: Sequelize.STRING,
-    score: Sequelize.INTEGER,
+    score: Sequelize.FLOAT,
     create_time: Sequelize.BIGINT,
     update_time: Sequelize.BIGINT,
     delete_time: Sequelize.BIGINT,
